Leverage transformCursor
