+++ lib/json0-presence.js
@@ -1,3 +1,6 @@
+const { transformCursor } = require('./text0');
+const { convertFromText } = require('./utils');
+
 // Draws from https://github.com/Teamwork/ot-rich-text/blob/master/src/Operation.js
 function isValidPresence(presence) {
   if (
@@ -18,7 +21,7 @@ function isValidPresence(presence) {
 
     if (
       !Array.isArray(selection) ||
-      selection.length !== 2 ||
+      selection.length < 2 ||
       selection[0] !== (selection[0] | 0) ||
       selection[1] !== (selection[1] | 0)
     ) {
@@ -53,33 +56,19 @@ const comparePresence = function(pres1, pres2) {
   } else return pres1 === pres2;
 };
 
-// this is the key function, always run client-side, both on
-// the client that creates a text-change, and on the clients
-// that receive text-changes (ops). if there are no ops, just
-// return presence, if there are ops, delegate to the subtype
-// responsible for those ops (currently only ot-rich-text).
-// I am making assumptions many places that all ops will be
-// of the same subtype, not sure if this is a given.
-// We're only concerned about the first level of object/array,
-// not sure if the spec allows nesting of subtypes.
-const transformPresence = function(presence, op, isOwn) {
-  if (op.length < 1) {
-    return presence;
-  }
-  const representativeOp = op[0];
-  const opType = op[0].t;
-  const path = representativeOp.p && representativeOp.p[0];
-  if (opType && subtypes[opType] && path) {
-    if (!presence.p || !presence.p[0] || presence.p[0] !== path) {
-      return presence;
+const transformPresence = (presence, op, isOwn) => {
+  for (let c of op){
+    if(c.si || c.sd){
+      convertFromText(c)
     }
-    // return result of running the subtype's transformPresence,
-    // but add path and type, which the subtype will not include
-    presence = {
-      ...subtypes[opType].transformPresence(presence, op, isOwn),
-      p: op[0].p,
-      t: op[0].t,
-    };
+    if(c.t === 'text0') {
+      presence = Object.assign({}, presence, {
+        s: presence.s.map(([start, end]) => ([
+          transformCursor(start, c.o),
+          transformCursor(end, c.o),
+        ]))
+      });
+    }
   }
   return presence;
 };
