+++ lib/json0.js
@@ -1,4 +1,7 @@
 var presence = require('./json0-presence')
+var utils = require('./utils');
+var convertFromText = utils.convertFromText;
+var convertToText = utils.convertToText;
 /*
  This is the implementation of the JSON OT type.
 
@@ -116,23 +119,6 @@ json.checkObj = function(elem) {
     throw new Error("Referenced element not an object (it was " + JSON.stringify(elem) + ")");
   }
 };
-
-// helper functions to convert old string ops to and from subtype ops
-function convertFromText(c) {
-  c.t = 'text0';
-  var o = {p: c.p.pop()};
-  if (c.si != null) o.i = c.si;
-  if (c.sd != null) o.d = c.sd;
-  c.o = [o];
-}
-
-function convertToText(c) {
-  c.p.push(c.o[0].p);
-  if (c.o[0].i != null) c.si = c.o[0].i;
-  if (c.o[0].d != null) c.sd = c.o[0].d;
-  delete c.t;
-  delete c.o;
-}
 
 json.apply = function(snapshot, op) {
   json.checkValidOp(op);
